Fixed file checks (wget still created an empty file) Refactored

+++ extras/deb_package_files/home/kootnet_data/scripts/update_kootnet-sensors_http.sh
@@ -5,7 +5,6 @@ if [[ $EUID != 0 ]]; then
   sudo "$0" "$@"
   exit $?
 fi
-clear
 HTTP_SERVER="http://kootenay-networks.com"
 HTTP_FOLDER="/installers"
 DEB_INSTALLER="/KootnetSensors_online.deb"
@@ -16,12 +15,20 @@ if [[ "$1" == "dev" ]]; then
   HTTP_FOLDER="/installers/dev"
   INSTALL_TYPE="Developmental"
 fi
-printf '\n-- %s HTTP Upgrade or Install --\n\n' "${INSTALL_TYPE}"
+clear
+printf '%s HTTP Upgrade or Install\n\n' "${INSTALL_TYPE}"
+printf 'Downloading installer ...\n\n'
 # Clean up previous downloads if any
 rm -f /tmp${DEB_INSTALLER} 2>/dev/null
 wget -O /tmp${DEB_INSTALLER} ${HTTP_SERVER}${HTTP_FOLDER}${DEB_INSTALLER}
-apt-get update
-apt-get -y install /tmp${DEB_INSTALLER}
-# Save DateTime and Update type to file (Used in program to show last updated)
-date -u >${CONFIG_DIR}/last_updated.txt
-echo ' - HTTP ' >>${CONFIG_DIR}/last_updated.txt
+# Make sure the installer file is there
+if [[ -s /tmp${DEB_INSTALLER} ]]; then
+  printf 'Download complete\nStarting Upgrade\n\n'
+  apt-get update
+  apt-get -y install /tmp${DEB_INSTALLER}
+  # Save DateTime and Update type to file (Used in program to show last updated)
+  date -u >${CONFIG_DIR}/last_updated.txt
+  echo ' - HTTP ' >>${CONFIG_DIR}/last_updated.txt
+else
+  printf '\nDownload failed, %s HTTP upgrade cancelled\n\n' "${INSTALL_TYPE}"
+fi
